+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Version 0.2
+
+* Layout extended to show two tables (search and queue) plus search field and
+  playback buttons
+* Search field has a prefix to search for:
+  * `/1` artist
+  * `/2` album
+  * `/3` track
+  Hitting return executes the search.
+* Double click or return in queue plays song
+* Double click or return in search appends the song or album to the queue
+
 ## Version 0.1
 
 * A first prototype to connect to a device and play a song from queue via double